Extract arithmetic instructions to instructions_arithmetic.go module
